Moved interactor factory types into interactors.json
