+++ Interactors/Hardware/GetHardwareDetailsInteractor.py
